+++ src/strongbox/core.clj
@@ -153,6 +153,9 @@
    ;; the list of addons from the catalogue
    :db nil
 
+   ;; some generated stats about the db that are updated just once at load time.
+   :db-stats {:known-host-list []}
+
    ;; the list of addons from the user-catalogue
    :user-catalogue nil
 
@@ -864,8 +867,14 @@
                              (load-current-catalogue))]
       (when-not (empty? final-catalogue)
         (p :p2/db:load
-           (swap! state assoc :db
-                  (db/put-many [] (:addon-summary-list final-catalogue))))))
+           (swap! state merge {:db (:addon-summary-list final-catalogue)
+                               :db-stats {:num-addons (count (:addon-summary-list final-catalogue))
+                                          :known-host-list (->> final-catalogue
+                                                                :addon-summary-list
+                                                                (map :source)
+                                                                distinct
+                                                                sort
+                                                                vec)}}))))
     (debug "skipping db load. already loaded or no catalogue selected."))
   nil)